Move keyboard layout logic into the .kv file.
I wish I knew a better way to do this. What’s in the .kv file now is
gross WET code but it saves having to deal with binding everything up
in the Python to change the color of a key. And a keyboard layout is
really layout logic, and is pretty static.