Highlight the first note of a Beastie attack.
I don’t like how this is done, but it works. Creatures shouldn’t have
to know about the screen they’re on. It would probably suffice to link
to the Keyboard instead, or even have Beasties dispatch events to the
party they’re in instead of the Screen instance. Or both.